Finishing PK code clean-up: thread workers done.
The thread workers manage integral storage and give each thread
the shell quartets it should compute when necessary.
They handle AIO to the various files for PK.

Code needs debugging and testing. Might not compile.
