+++ src/lib/libfock/writers.cc
@@ -270,1644 +270,5 @@ void IWLAsync::flush(int lastbuf) {
     idx_ = 0;
 }
 
-// IWLAsync_PK class implementation
-
-IWLAsync_PK::IWLAsync_PK(size_t *address, boost::shared_ptr<AIOHandler> AIO, int itap) {
-    itap_ = itap;
-    address_ = address;
-    AIO_ = AIO;
-    ints_per_buf_ = IWL_INTS_PER_BUF;
-    nints_ = 0;
-    idx_ = 0;
-    labels_[0] = new Label[4 * ints_per_buf_];
-    labels_[1] = new Label[4 * ints_per_buf_];
-    values_[0] = new Value[ints_per_buf_];
-    values_[1] = new Value[ints_per_buf_];
-    JobID_[0] = 0;
-    JobID_[1] = 0;
-    lastbuf_ = 0;
-}
-
-IWLAsync_PK::~IWLAsync_PK() {
-    delete [] labels_[0];
-    delete [] labels_[1];
-    delete [] values_[0];
-    delete [] values_[1];
-}
-
-void IWLAsync_PK::fill_values(double val, size_t i, size_t j, size_t k, size_t l) {
-    labels_[idx_][4 * nints_] = i;
-    labels_[idx_][4 * nints_ + 1] = j;
-    labels_[idx_][4 * nints_ + 2] = k;
-    labels_[idx_][4 * nints_ + 3] = l;
-    values_[idx_][nints_] = val;
-    ++nints_;
-}
-
-// Buffer is full, write it using AIO to disk
-void IWLAsync_PK::write() {
-    size_t lab_size = 4 * ints_per_buf_ * sizeof(Label);
-    size_t val_size = ints_per_buf_ * sizeof(Value);
-    // We need a special function in AIO to take care
-    // of IWL buffer writing, since these contain four parts.
-    JobID_[idx_] = AIO_->write_iwl(itap_,IWL_KEY_BUF,nints_,lastbuf_,(char *) labels_[idx_],
-                    (char*) values_[idx_], lab_size, val_size, address_);
-
-    // Now we need to switch the internal buffer to which we are writing.
-    idx_ = idx_ == 0 ? 1 : 0;
-    nints_ = 0;
-    AIO_->wait_for_job(JobID_[idx_]);
-}
-
-// Pop a value from the buffer storage
-void IWLAsync_PK::pop_value(double &val, size_t &i, size_t &j, size_t &k, size_t &l) {
-    if(nints_ == 0) {
-        throw PSIEXCEPTION("Cannot pop value from empty buffer\n");
-    }
-    --nints_;
-    i = labels_[idx_][4 * nints_];
-    j = labels_[idx_][4 * nints_ + 1];
-    k = labels_[idx_][4 * nints_ + 2];
-    l = labels_[idx_][4 * nints_ + 3];
-    val = values_[idx_][nints_];
-}
-
-void IWLAsync_PK::flush() {
-    unsigned int nints = nints_;
-    while(nints_ != ints_per_buf_) {
-        fill_values(0.0,0,0,0,0);
-    }
-    nints_ = nints;
-    lastbuf_ = 1;
-    write();
-}
-
-void ijklBasisIterator::first() {
-    i_ = 0;
-    j_ = 0;
-    k_ = 0;
-    l_ = 0;
-}
-
-void ijklBasisIterator::next() {
-    ++l_;
-    if (l_ > j_ && k_ == i_ ) {
-        l_ = 0;
-        ++k_;
-    }
-    if (l_ > k_) {
-        l_ = 0;
-        ++k_;
-    }
-    if (k_ > i_) {
-        k_ = 0;
-        ++j_;
-        if (j_ > i_) {
-            j_ = 0;
-            ++i_;
-            if (i_ >= nbas_) {
-                done_=true;
-            }
-        }
-    }
-}
-
-IOBuffer_PK::IOBuffer_PK(boost::shared_ptr<BasisSet> primary, boost::shared_ptr<AIOHandler> AIO,
-                         size_t buf_size, size_t nbuf, int pk_file) {
-    AIO_ = AIO;
-    pk_file_ = pk_file;
-    primary_ = primary;
-    buf_size_ = buf_size;
-    nbuf_ = nbuf;
-    offset_ = 0;
-    bufidx_ = 0;
-    buf_ = 0;
-    allocate();
-}
-
-IOBuffer_PK::IOBuffer_PK(boost::shared_ptr<BasisSet> primary, size_t buf_size) {
-    primary_ = primary;
-    buf_size_ = buf_size;
-    offset_ = 0;
-    bufidx_ = 0;
-    buf_ = 0;
-}
-
-IOBuffer_PK::~IOBuffer_PK() {
-    deallocate();
-}
-
-void IOBuffer_PK::allocate() {
-    for (unsigned int i = 0; i < nbuf_; ++i) {
-        double* buf = new double[buf_size_];
-        J_buf_.push_back(buf);
-        buf = new double[buf_size_];
-        K_buf_.push_back(buf);
-        std::vector<char*> label_J;
-        std::vector<char*> label_K;
-        label_J_.push_back(label_J);
-        label_K_.push_back(label_K);
-        std::vector<size_t> id_J;
-        std::vector<size_t> id_K;
-        jobID_J_.push_back(id_J);
-        jobID_K_.push_back(id_K);
-    }
-    // We only need to set to zero the first buffers.
-    // The others are taken care of in the write function
-    ::memset((void*) J_buf_[0], '\0', buf_size_ * sizeof(double));
-    ::memset((void*) K_buf_[0], '\0', buf_size_ * sizeof(double));
-}
-
-void IOBuffer_PK::deallocate() {
-    std::vector<double*>::iterator it;
-    for(it = J_buf_.begin(); it != J_buf_.end(); ++it) {
-        delete [] *it;
-    }
-    J_buf_.clear();
-    for(it = K_buf_.begin(); it != K_buf_.end(); ++it) {
-        delete [] *it;
-    }
-    K_buf_.clear();
-    for(int i = 0; i < label_J_.size(); ++i) {
-        for(int j = 0; j < label_J_[i].size(); ++j) {
-            delete [] label_J_[i][j];
-        }
-        label_J_[i].clear();
-    }
-    for(int i = 0; i < label_K_.size(); ++i) {
-        for(int j = 0; j < label_K_[i].size(); ++j) {
-            delete [] label_K_[i][j];
-        }
-        label_K_[i].clear();
-    }
-}
-
-void IOBuffer_PK::first_quartet(size_t i) {
-    shelliter_ = AOShellCombinationsIterator(primary_,primary_,primary_,primary_);
-    bufidx_ = i;
-    offset_ = bufidx_ * buf_size_;
-    shells_left_ = false;
-    for(shelliter_.first(); (shells_left_ || shelliter_.is_done()) == false; shelliter_.next()) {
-        P_ = shelliter_.p();
-        Q_ = shelliter_.q();
-        R_ = shelliter_.r();
-        S_ = shelliter_.s();
-
-        shells_left_ = is_shell_relevant();
-    }
-
-}
-
-void IOBuffer_PK::next_quartet() {
-    if(shelliter_.is_done()) {
-        shells_left_ = false;
-        return;
-    }
-    bool shell_found = false;
-    for(; (shell_found || shelliter_.is_done()) == false; shelliter_.next()) {
-        P_ = shelliter_.p();
-        Q_ = shelliter_.q();
-        R_ = shelliter_.r();
-        S_ = shelliter_.s();
-
-        shell_found = is_shell_relevant();
-    }
-
-    shells_left_ = shell_found;
-}
-
-bool IOBuffer_PK::is_shell_relevant() {
-        // May implement the sieve here
-
-        size_t lowi = primary_->shell_to_basis_function(P_);
-        size_t lowj = primary_->shell_to_basis_function(Q_);
-        size_t lowk = primary_->shell_to_basis_function(R_);
-        size_t lowl = primary_->shell_to_basis_function(S_);
-
-        size_t low_ijkl = INDEX4(lowi, lowj, lowk, lowl);
-        size_t low_ikjl = INDEX4(lowi, lowk, lowj, lowl);
-        size_t low_iljk = INDEX4(lowi, lowl, lowj, lowk);
-
-        size_t nbJ_lo = low_ijkl / buf_size_;
-        size_t nbK1_lo = low_ikjl / buf_size_;
-        size_t nbK2_lo = low_iljk / buf_size_;
-
-        // Can we filter out this shell because all of its basis function
-        // indices are too high ?
-
-        if (nbJ_lo > bufidx_ && nbK1_lo > bufidx_ && nbK2_lo > bufidx_) {
-            return false;
-        }
-
-        int ni = primary_->shell(P_).nfunction();
-        int nj = primary_->shell(Q_).nfunction();
-        int nk = primary_->shell(R_).nfunction();
-        int nl = primary_->shell(S_).nfunction();
-
-        size_t hii = lowi + ni - 1;
-        size_t hij = lowj + nj - 1;
-        size_t hik = lowk + nk - 1;
-        size_t hil = lowl + nl - 1;
-
-        size_t hi_ijkl = INDEX4(hii, hij, hik, hil);
-        size_t hi_ikjl = INDEX4(hii, hik, hij, hil);
-        size_t hi_iljk = INDEX4(hii, hil, hij, hik);
-
-        size_t nbJ_hi = hi_ijkl / buf_size_;
-        size_t nbK1_hi = hi_ikjl / buf_size_;
-        size_t nbK2_hi = hi_iljk / buf_size_;
-
-        // Can we filter out this shell because all of its basis function
-        // indices are too low ?
-
-        if (nbJ_hi < bufidx_ && nbK1_hi < bufidx_ && nbK2_hi < bufidx_) {
-            return false;
-        }
-
-        // Now we loop over unique basis function quartets in the shell quartet
-        AOIntegralsIterator bfiter = shelliter_.integrals_iterator();
-        for(bfiter.first(); bfiter.is_done() == false; bfiter.next()) {
-            size_t i = bfiter.i();
-            size_t j = bfiter.j();
-            size_t k = bfiter.k();
-            size_t l = bfiter.l();
-
-            size_t ijkl = INDEX4(i,j,k,l);
-            size_t ikjl = INDEX4(i,k,j,l);
-            size_t iljk = INDEX4(i,l,j,k);
-
-            size_t nbJ = ijkl / buf_size_;
-            size_t nbK1 = ikjl / buf_size_;
-            size_t nbK2 = iljk / buf_size_;
-
-            if (nbJ == bufidx_ || nbK1 == bufidx_ || nbK2 == bufidx_) {
-                // This shell should be computed by the present thread.
-                return true;
-            }
-        }
-        return false;
-}
-
-void IOBuffer_PK::fill_values(double val, size_t i, size_t j, size_t k, size_t l) {
-    size_t ijkl = INDEX4(i,j,k,l);
-    if (ijkl / buf_size_ == bufidx_) {
-        J_buf_[buf_][ijkl - offset_] += val;
-    }
-    size_t ikjl = INDEX4(i, k, j, l);
-    if(ikjl / buf_size_ == bufidx_) {
-        if (i == k || j == l) {
-            K_buf_[buf_][ikjl - offset_] += val;
-        } else {
-            K_buf_[buf_][ikjl - offset_] += 0.5 * val;
-        }
-    }
-
-    if(i != j && k != l) {
-        size_t iljk = INDEX4(i, l, j, k);
-        if (iljk / buf_size_ == bufidx_) {
-            if ( i == l || j == k) {
-                K_buf_[buf_][iljk - offset_] += val;
-            } else {
-                K_buf_[buf_][iljk - offset_] += 0.5 * val;
-            }
-        }
-    }
-}
-
-// Pretty sure I'm leaking memory over there
-void IOBuffer_PK::write(std::vector<size_t> &batch_min_ind, std::vector<size_t> &batch_max_ind,
-                   size_t pk_pairs) {
-    // Compute initial ijkl index for current buffer
-    size_t ijkl0 = buf_size_ * bufidx_;
-    size_t ijkl_end = ijkl0 + buf_size_ - 1;
-
-    std::vector<unsigned int> target_batches;  // Vector of batch number to which we want to write
-
-    for (unsigned int i = 0; i < batch_min_ind.size(); ++i) {
-        if (ijkl0 >= batch_min_ind[i] && ijkl0 < batch_max_ind[i]) {
-            target_batches.push_back(i);
-            continue;
-        }
-        if (ijkl_end >= batch_min_ind[i] && ijkl_end < batch_max_ind[i]) {
-            target_batches.push_back(i);
-            continue;
-        }
-        if (ijkl0 < batch_min_ind[i] && ijkl_end >= batch_max_ind[i]) {
-            target_batches.push_back(i);
-            continue;
-        }
-    }
-
-    // Now that all buffers are full, and before we write integrals, we need
-    // to divide diagonal elements by 2.
-
-    for (size_t pq = 0; pq < pk_pairs; ++pq) {
-        size_t pqpq = INDEX2(pq, pq);
-        if (pqpq >= ijkl0 && pqpq <= ijkl_end) {
-            J_buf_[buf_][pqpq - offset_] *= 0.5;
-            K_buf_[buf_][pqpq - offset_] *= 0.5;
-        }
-    }
-
-    // And now we write to the file in the appropriate entries
-    for (int i = 0; i < target_batches.size(); ++i) {
-        label_J_[buf_].push_back(PK_integrals_old::get_label_J(target_batches[i]));
-        size_t start = std::max(ijkl0, batch_min_ind[target_batches[i]]);
-        size_t stop = std::min(ijkl_end + 1, batch_max_ind[target_batches[i]]);
-        psio_address adr = psio_get_address(PSIO_ZERO, (start - batch_min_ind[target_batches[i]]) * sizeof(double));
-        size_t nints = stop - start;
-        jobID_J_[buf_].push_back(AIO_->write(pk_file_, label_J_[buf_][i], (char *)(&J_buf_[buf_][start - offset_]),
-                    nints * sizeof(double), adr, &dummy_));
-        label_K_[buf_].push_back(PK_integrals_old::get_label_K(target_batches[i]));
-        jobID_K_[buf_].push_back(AIO_->write(pk_file_, label_K_[buf_][i], (char *)(&K_buf_[buf_][start - offset_]),
-                    nints * sizeof(double), adr, &dummy_));
-    }
-
-    // Update the buffer being written into
-    ++buf_;
-    if (buf_ >= nbuf_) buf_ = 0;
-    // Make sure the buffer has been written to disk and we can erase it
-    for(int i = 0; i < jobID_J_[buf_].size(); ++i) {
-      AIO_->wait_for_job(jobID_J_[buf_][i]);
-    }
-    jobID_J_[buf_].clear();
-    for(int i = 0; i < jobID_K_[buf_].size(); ++i) {
-      AIO_->wait_for_job(jobID_K_[buf_][i]);
-    }
-    jobID_K_[buf_].clear();
-    // We can delete the labels for these buffers
-    for(int i = 0; i < label_J_[buf_].size(); ++i) {
-        delete [] label_J_[buf_][i];
-    }
-    for(int i = 0; i < label_K_[buf_].size(); ++i) {
-        delete [] label_K_[buf_][i];
-    }
-    label_J_[buf_].clear();
-    label_K_[buf_].clear();
-    // Make sure the buffer in which we are going to write is set to zero
-    ::memset((void *) J_buf_[buf_], '\0', buf_size_ * sizeof(double));
-    ::memset((void *) K_buf_[buf_], '\0', buf_size_ * sizeof(double));
-}
-
-InCoreBufferPK::InCoreBufferPK(boost::shared_ptr<BasisSet> primary, size_t buf_size, size_t lastbuf,
-                               double *Jbuf, double *Kbuf) : IOBuffer_PK(primary,buf_size) {
-    last_buf_ = lastbuf;
-    J_bufp_ = Jbuf;
-    K_bufp_ = Kbuf;
-}
-
-bool InCoreBufferPK::is_shell_relevant() {
-        // May implement the sieve here
-
-        size_t lowi = primary_->shell_to_basis_function(P_);
-        size_t lowj = primary_->shell_to_basis_function(Q_);
-        size_t lowk = primary_->shell_to_basis_function(R_);
-        size_t lowl = primary_->shell_to_basis_function(S_);
-
-        size_t low_ijkl = INDEX4(lowi, lowj, lowk, lowl);
-        size_t low_ikjl = INDEX4(lowi, lowk, lowj, lowl);
-        size_t low_iljk = INDEX4(lowi, lowl, lowj, lowk);
-
-        // Use the max ijkl that can be in the buffer
-        size_t max_idx = (buf_size_ * (bufidx_ + 1) + last_buf_) - 1;
-        //outfile->Printf("max idx is %lu\n",max_idx);
-
-        // Can we filter out this shell because all of its basis function
-        // indices are too high ?
-
-        if (low_ijkl > max_idx && low_ikjl > max_idx && low_iljk > max_idx) {
-            return false;
-        }
-
-        int ni = primary_->shell(P_).nfunction();
-        int nj = primary_->shell(Q_).nfunction();
-        int nk = primary_->shell(R_).nfunction();
-        int nl = primary_->shell(S_).nfunction();
-
-        size_t hii = lowi + ni - 1;
-        size_t hij = lowj + nj - 1;
-        size_t hik = lowk + nk - 1;
-        size_t hil = lowl + nl - 1;
-
-        size_t hi_ijkl = INDEX4(hii, hij, hik, hil);
-        size_t hi_ikjl = INDEX4(hii, hik, hij, hil);
-        size_t hi_iljk = INDEX4(hii, hil, hij, hik);
-
-        // Use the min ijkl that can be in buffer, which is the offset_
-
-        // Can we filter out this shell because all of its basis function
-        // indices are too low ?
-
-        if (hi_ijkl < offset_ && hi_ikjl < offset_ && hi_iljk < offset_) {
-            return false;
-        }
-
-        // Now we loop over unique basis function quartets in the shell quartet
-        AOIntegralsIterator bfiter = shelliter_.integrals_iterator();
-        for(bfiter.first(); bfiter.is_done() == false; bfiter.next()) {
-            size_t i = bfiter.i();
-            size_t j = bfiter.j();
-            size_t k = bfiter.k();
-            size_t l = bfiter.l();
-
-            size_t ijkl = INDEX4(i,j,k,l);
-            size_t ikjl = INDEX4(i,k,j,l);
-            size_t iljk = INDEX4(i,l,j,k);
-
-            bool bJ = ijkl >= offset_ && ijkl <= max_idx;
-            bool bK1 = ikjl >= offset_ && ikjl <= max_idx;
-            bool bK2 = iljk >= offset_ && iljk <= max_idx;
-
-            if (bJ || bK1 || bK2) {
-                // This shell should be computed by the present thread.
-                return true;
-            }
-        }
-
-        return false;
-
-}
-
-void InCoreBufferPK::fill_values(double val, size_t i, size_t j, size_t k, size_t l) {
-    size_t ijkl = INDEX4(i,j,k,l);
-    size_t ikjl = INDEX4(i, k, j, l);
-    // Use the max ijkl that can be in the buffer
-    size_t max_idx = (buf_size_ * (bufidx_ + 1) + last_buf_) - 1;
-    //    outfile->Printf("max idx is %lu\n",max_idx);
-
-    if (ijkl >= offset_ && ijkl <= max_idx) {
-        J_bufp_[ijkl - offset_] += val;
-    }
-    if(ikjl >= offset_ && ikjl <= max_idx) {
-        if (i == k || j == l) {
-            K_bufp_[ikjl - offset_] += val;
-        } else {
-            K_bufp_[ikjl - offset_] += 0.5 * val;
-        }
-    }
-
-    if(i != j && k != l) {
-        size_t iljk = INDEX4(i, l, j, k);
-        if (iljk >= offset_ && iljk <= max_idx) {
-            if ( i == l || j == k) {
-                K_bufp_[iljk - offset_] += val;
-            } else {
-                K_bufp_[iljk - offset_] += 0.5 * val;
-            }
-        }
-    }
-
-}
-
-// Note: this function does not actually write anything, it divides by 2 diagonal
-// elements of the array once we know they are all computed
-
-void InCoreBufferPK::write(std::vector<size_t> &batch_min_ind, std::vector<size_t> &batch_max_ind, size_t pk_pairs) {
-    // And we do not actually use all arguments
-
-    size_t max_idx = (buf_size_ * (bufidx_ + 1) + last_buf_) - 1;
-        outfile->Printf("max idx is %lu\n",max_idx);
-    for (size_t pq = 0; pq < pk_pairs; ++pq) {
-        size_t pqpq = INDEX2(pq,pq);
-        if(pqpq >= offset_ && pqpq <= max_idx) {
-            J_bufp_[pqpq - offset_] *= 0.5;
-            K_bufp_[pqpq - offset_] *= 0.5;
-        }
-    }
-}
-
-IOBuffer_IWL::IOBuffer_IWL(boost::shared_ptr<BasisSet> primary,
-                           boost::shared_ptr<AIOHandler> AIO, size_t nbuf, size_t buf_size,
-                           int J_file, int K_file, size_t *pos, std::vector<int> *bufforpq)
-    : IOBuffer_PK(primary,buf_size) {
-
-    pk_file_ = J_file;
-    nbuf_ = nbuf;
-    AIO_ = AIO;
-    K_file_ = K_file;
-    addresses_ = pos;
-    buf_for_pq_ = bufforpq;
-    allocate();
-
-}
-
-void IOBuffer_IWL::allocate() {
-    for(int i = 0; i < nbuf_; ++i) {
-        bufs_J_.push_back(new IWLAsync_PK(&addresses_[2 * i],AIO_,pk_file_));
-        bufs_K_.push_back(new IWLAsync_PK(&addresses_[2 * i + 1],AIO_,K_file_));
-    }
-}
-
-void IOBuffer_IWL::deallocate() {
-    for(int i = 0; i < nbuf_; ++i) {
-        delete bufs_J_[i];
-        delete bufs_K_[i];
-    }
-}
-
-void IOBuffer_IWL::fill_values(double val, size_t i, size_t j, size_t k, size_t l) {
-    // Pre-sorting for J
-    size_t pq = INDEX2(i,j);
-    IWLAsync_PK* buf = bufs_J_[buf_for_pq_->at(pq)];
-    buf->fill_values(val,i,j,k,l);
-    if(buf->nints() == buf->maxints()) {
-        buf->write();
-    }
-
-    // Pre-sorting for K
-    pq = INDEX2(i,k);
-    int bufK1 = buf_for_pq_->at(pq);
-    buf = bufs_K_[bufK1];
-    buf->fill_values(val,i,j,k,l);
-    if(buf->nints() == buf->maxints()) {
-        buf->write();
-    }
-    // Second pre-sorting for K
-    if (i != j && k != l) {
-        pq = std::max(INDEX2(i,l), INDEX2(j,k));
-        int bufK2 = buf_for_pq_->at(pq);
-        if (bufK2 != bufK1) {
-            buf = bufs_K_[bufK2];
-            buf->fill_values(val,i,j,k,l);
-            if(buf->nints() == buf->maxints()) {
-                buf->write();
-            }
-        }
-    }
-}
-
-bool IOBuffer_IWL::pop_value(unsigned int bufid, double &val, size_t &i, size_t &j, size_t &k, size_t &l) {
-    IWLAsync_PK* buf;
-    if(bufid < nbuf_) {
-        buf = bufs_J_[bufid];
-    } else {
-        buf = bufs_K_[bufid - nbuf_];
-    }
-    if(buf->nints() == 0) {
-        return false;
-    }
-    buf->pop_value(val,i,j,k,l);
-    return true;
-}
-
-void IOBuffer_IWL::insert_value(unsigned int bufid, double val, size_t i, size_t j, size_t k, size_t l) {
-    IWLAsync_PK* buf;
-    if(bufid < nbuf_) {
-        buf = bufs_J_[bufid];
-    } else {
-        buf = bufs_K_[bufid - nbuf_];
-    }
-    buf->fill_values(val,i,j,k,l);
-}
-
-void IOBuffer_IWL::flush() {
-    IWLAsync_PK* buf;
-    for(int bufid = 0; bufid < nbuf_; ++bufid) {
-        buf = bufs_J_[bufid];
-        buf->flush();
-        buf = bufs_K_[bufid];
-        buf->flush();
-    }
-}
-
-PK_integrals_old::PK_integrals_old(boost::shared_ptr<BasisSet> primary, boost::shared_ptr<PSIO> psio,
-                           int max_batches, size_t memory, double cutoff) {
-    primary_ = primary;
-    nbf_ = primary_->nbf();
-    max_batches_ = max_batches;
-    memory_ = memory;
-
-    pk_pairs_ = nbf_ * (nbf_ + 1) / 2;
-    pk_size_ = pk_pairs_ * (pk_pairs_ + 1) / 2;
-    if (2 * pk_size_ < memory_ ) {
-        in_core_ = true;
-    } else {
-        in_core_ = false;
-    }
-
-    if (memory_ < pk_pairs_) {
-        throw PSIEXCEPTION("Not enough memory for PK algorithm");
-    }
-
-    J_buf_[0] = NULL;
-    J_buf_[1] = NULL;
-    K_buf_[0] = NULL;
-    K_buf_[1] = NULL;
-    sieve_ = boost::shared_ptr<ERISieve>(new ERISieve(primary_, cutoff));
-
-    nbuffers_ = 0;
-
-    psio_ = psio;
-    AIO_ = boost::shared_ptr<AIOHandler>(new AIOHandler(psio_));
-
-    itap_J_ = PSIF_SO_PKSUPER1;
-    itap_K_ = PSIF_SO_PKSUPER2;
-
-}
-
-PK_integrals::PK_integrals(boost::shared_ptr<BasisSet> primary, boost::shared_ptr<PSIO> psio, int max_batches,
-                           size_t memory, double cutoff) :
-        PK_integrals_old(primary,psio, max_batches, memory, cutoff) {
-    iwl_file_J_ = PSIF_SO_TEI;
-    iwl_file_K_ = PSIF_SO_PKSUPER1;
-    pk_file_ = PSIF_SO_PK;
-    nthreads_ = 1;
-#ifdef _OPENMP
-    nthreads_ = omp_get_max_threads();
-#endif
-    tgt_tasks_ = 100;
-    max_mem_buf_ = 71520;
-    max_mem_buf_ = Process::environment.options.get_int("MAX_MEM_BUF");
-    itap_J_ = pk_file_;
-    itap_K_ = pk_file_;
-    ints_per_buf_ = IWL_INTS_PER_BUF;
-    iwlintsize_ = 2L * sizeof(int) + ints_per_buf_ *
-            (4L * sizeof(Label) + sizeof(Value));
-
-}
-
-PK_integrals_old::~PK_integrals_old() {
-    // In case we forgot something before
-
-    if (J_buf_[0] != NULL) {
-        delete [] J_buf_[0];
-        J_buf_[0] = NULL;
-        outfile->Printf("Clean up your code!!!");
-    }
-    if (J_buf_[1] != NULL) {
-        delete [] J_buf_[1];
-        J_buf_[1] = NULL;
-        outfile->Printf("Clean up your code!!!");
-    }
-    if (K_buf_[0] != NULL) {
-        delete [] K_buf_[0];
-        K_buf_[0] = NULL;
-        outfile->Printf("Clean up your code!!!");
-    }
-    if (K_buf_[1] != NULL) {
-        delete [] K_buf_[1];
-        K_buf_[1] = NULL;
-        outfile->Printf("Clean up your code!!!");
-    }
-    if (label_J_[0].size() != 0) {
-        outfile->Printf("Labels: Clean up your code!!!");
-    }
-    if (label_J_[1].size() != 0) {
-        outfile->Printf("Labels: Clean up your code!!!");
-    }
-    if (label_K_[0].size() != 0) {
-        outfile->Printf("Labels: Clean up your code!!!");
-    }
-    if (label_K_[1].size() != 0) {
-        outfile->Printf("Labels: Clean up your code!!!");
-    }
-
-    //If we did in-core, now is the time to delete the big buffers
-    if (in_core_) {
-        delete [] J_ints_;
-        delete [] K_ints_;
-    }
-}
-
-void PK_integrals_old::batch_sizing() {
-
-    if (in_core_) {
-        return;
-    }
-
-    double batch_thresh = 0.1;
-
-    ijklBasisIterator AOintsiter(nbf_);
-
-    size_t old_pq = 0;
-    size_t old_max = 0;
-    size_t nintpq = 0;
-    size_t nintbatch = 0;
-    size_t pq = 0;
-    size_t pb, qb, rb, sb;
-    int batch = 0;
-
-    batch_index_min_.push_back(0);
-    batch_pq_min_.push_back(0);
-    batch_for_pq_.push_back(0);
-    for (AOintsiter.first(); AOintsiter.is_done() == false; AOintsiter.next()) {
-        pb = AOintsiter.i();
-        qb = AOintsiter.j();
-        rb = AOintsiter.k();
-        sb = AOintsiter.l();
-
-        pq = INDEX2(pb, qb);
-
-        if (old_pq == pq) {
-            ++nintpq;
-        } else {
-            size_t pqrs = INDEX2(pq, INDEX2(rb, sb));
-            nintbatch += nintpq;
-            if (nintbatch > memory_) {
-                batch_index_max_.push_back(old_max);
-                batch_pq_max_.push_back(old_pq);
-                batch_for_pq_.pop_back();
-                ++batch;
-                batch_for_pq_.push_back(batch);
-                batch_index_min_.push_back(old_max);
-                batch_pq_min_.push_back(old_pq);
-                nintbatch = nintpq;
-            }
-            nintpq = 1;
-            old_pq = pq;
-            batch_for_pq_.push_back(batch);
-            old_max = pqrs;
-        }
-    }
-    batch_index_max_.push_back(INDEX4(pb,qb,rb,sb) + 1);
-    batch_pq_max_.push_back(INDEX2(pb,qb) + 1);
-
-    // A little check here: if the last batch is less than 10% full,
-    // we just transfer it to the previous batch.
-
-    int lastb = batch_index_max_.size() - 1;
-    if (lastb > 0) {
-    size_t size_lastb = batch_index_max_[lastb] - batch_index_min_[lastb];
-        if (((double) size_lastb / memory_) < batch_thresh) {
-            batch_index_max_[lastb - 1] = batch_index_max_[lastb];
-            batch_pq_max_[lastb - 1] = batch_pq_max_[lastb];
-            batch_pq_max_.pop_back();
-            batch_index_max_.pop_back();
-            batch_pq_min_.pop_back();
-            batch_index_min_.pop_back();
-        }
-    }
-
-    int nbatches = batch_pq_min_.size();
-    if (nbatches > max_batches_) {
-      outfile->Printf( "PKJK: maximum number of batches exceeded\n") ;
-      outfile->Printf( "   requested %d batches\n", nbatches) ;
-      exit(PSI_RETURN_FAILURE) ;
-    }
-
-}
-
-void PK_integrals_old::print_batches() {
-    if (in_core_) {
-        outfile->Printf("\tPerforming in-core PK\n");
-    } else {
-        // Print batches for the user and for control
-        for(int batch = 0; batch < batch_pq_min_.size(); ++batch){
-            outfile->Printf("\tBatch %3d pq = [%8zu,%8zu] index = [%14zu,%zu] size = %12zu\n",
-                    batch + 1,
-                    batch_pq_min_[batch],batch_pq_max_[batch],
-                    batch_index_min_[batch],batch_index_max_[batch],
-                    batch_index_max_[batch] - batch_index_min_[batch]);
-        }
-    }
-}
-
-void PK_integrals_old::allocate_buffers() {
-    buffer_size_ = memory_ / 4;
-    nbuffers_ = (pk_size_ - 1) / buffer_size_ + 1;
-    outfile->Printf("We need %3d buffers during integral computation\n",nbuffers_);
-    // Now we proceed to actually allocate buffer memory
-    J_buf_[0] = new double[buffer_size_];
-    J_buf_[1] = new double[buffer_size_];
-    K_buf_[0] = new double[buffer_size_];
-    K_buf_[1] = new double[buffer_size_];
-    ::memset((void*) J_buf_[0], '\0', buffer_size_ * sizeof(double));
-    ::memset((void*) J_buf_[1], '\0', buffer_size_ * sizeof(double));
-    ::memset((void*) K_buf_[0], '\0', buffer_size_ * sizeof(double));
-    ::memset((void*) K_buf_[1], '\0', buffer_size_ * sizeof(double));
-
-    bufidx_ = 0;
-    offset_ = 0;
-    buf_ = 0;
-
-    jobid_J_[0] = 0;
-    jobid_J_[1] = 0;
-    jobid_K_[0] = 0;
-    jobid_K_[1] = 0;
-}
-
-void PK_integrals::allocate_buffers() {
-    if(in_core_) {
-        // Need to allocate two big arrays
-        J_ints_ = new double[pk_size_];
-        K_ints_ = new double[pk_size_];
-        ::memset((void*) J_ints_, '\0', pk_size_ * sizeof(double));
-        ::memset((void*) K_ints_, '\0', pk_size_ * sizeof(double));
-
-        // Now we allocate a derived class of IOBuffer_PK that takes care of
-        // giving the tasks to the threads and storing results properly.
-
-        size_t buffer_size = pk_size_ / nthreads_;
-        size_t lastbuf = pk_size_ % nthreads_;
-        size_t start = 0;
-
-        for(int i = 0; i < nthreads_; ++i) {
-            start = i * buffer_size;
-            outfile->Printf("start is %lu\n",start);
-            IOBuffer_PK* buf;
-            if(i < nthreads_ - 1) {
-                buf = new InCoreBufferPK(primary_,buffer_size,0,&J_ints_[start],&K_ints_[start]);
-            } else {
-                buf = new InCoreBufferPK(primary_,buffer_size,lastbuf,&J_ints_[start],&K_ints_[start]);
-            }
-            iobuffers_.push_back(buf);
-            ntasks_ = nthreads_;
-        }
-
-    } else {
-        // Factor 2 because we need memory for J and K buffers
-        size_t mem_per_thread = memory_ / (2 * nthreads_);
-        // Factor 2 because we need 2 buffers for asynchronous I/O
-        size_t buf_size = mem_per_thread / 2;
-        // If there is only one buffer, we could use less memory
-        // In practice it should switch to in core PK
-        if (max_mem_buf_ != 0) buf_size = std::min(buf_size, max_mem_buf_);
-
-        // Number of tasks needed with this buffer size
-        ntasks_ = pk_size_ / buf_size + 1;
-        // Is there less tasks than threads ?
-        if(ntasks_ < nthreads_) {
-            ntasks_ = ntasks_ * nthreads_;
-            size_t tmp_size = pk_size_ / ntasks_ + 1;
-            if(tmp_size > buf_size) {
-                throw PSIEXCEPTION("I cannot math.\n");
-            }
-            buf_size = tmp_size;
-            ntasks_ = pk_size_ / buf_size + 1;
-        }
-        size_t buf_per_thread = std::min(mem_per_thread / buf_size, ntasks_ / nthreads_);
-        // Some printing for us
-        outfile->Printf("  Task number: %lu\n",ntasks_);
-        outfile->Printf("  Buffer size: %lu\n",buf_size);
-        outfile->Printf("  Buffer per thread: %lu\n",buf_per_thread);
-
-        // Ok, now we have the size of a buffer and how many buffers
-        // we want for each thread. We can allocate IO buffers.
-        for(int i = 0; i < nthreads_; ++i) {
-            iobuffers_.push_back(new IOBuffer_PK(primary_,AIO_,buf_size,buf_per_thread,pk_file_));
-        }
-    }
-}
-
-void PK_integrals::allocate_iwlbuffers() {
-    int buf_per_thread = batch_index_min_.size();
-    // Factor 2 because we need an address for J and an address for K
-    // J and K addresses for the same file are stored contiguously, i.e.
-    // element [0] is J address for first buffer and element [1] the
-    // K address for the first buffer.
-    size_t* current_pos = new size_t[2 * buf_per_thread];
-
-    current_pos[0] = 0;
-    current_pos[1] = 0;
-    //Current position of the bucket in the IWL file
-    // For K, each integral can potentially go into two buckets, so we
-    // give each bucket twice the J bucket's size. It's wasting quite a bit
-    // of space, other solution is to explicitly count integrals.
-    for(int i = 1; i < buf_per_thread; ++i) {
-        size_t batchsize = batch_index_max_[i - 1] - batch_index_min_[i - 1];
-        size_t iwlperbatch = batchsize / ints_per_buf_ + 1;
-        current_pos[2 * i] = iwlperbatch * iwlintsize_ + current_pos[2 * i - 2];
-        current_pos[2 * i + 1] = 2 * (iwlperbatch * iwlintsize_) + current_pos[2 * i - 1];
-    }
-
-    // Ok, now we have the size of a buffer and how many buffers
-    // we want for each thread. We can allocate IO buffers.
-    for(int i = 0; i < nthreads_; ++i) {
-        iobuffers_.push_back(new IOBuffer_IWL(primary_,AIO_,buf_per_thread,
-                                 ints_per_buf_,iwl_file_J_,iwl_file_K_,current_pos,&batch_for_pq_));
-    }
-}
-
-void PK_integrals::deallocate_iwlbuffers() {
-    for(int i = 0; i < nthreads_; ++i) {
-        delete iobuffers_[i];
-    }
-}
-
-IOBuffer_PK* PK_integrals::buffer() {
-    int thread = 0;
-#ifdef _OPENMP
-    thread = omp_get_thread_num();
-#endif
-    return iobuffers_[thread];
-}
-
-void PK_integrals_old::deallocate_buffers() {
-
-    // We need to make sure writing is over before deallocating!
-    timer_on("AIO synchronize");
-    AIO_->synchronize();
-    timer_off("AIO synchronize");
-
-    delete [] J_buf_[0];
-    delete [] J_buf_[1];
-    delete [] K_buf_[0];
-    delete [] K_buf_[1];
-
-    J_buf_[0] = NULL;
-    J_buf_[1] = NULL;
-    K_buf_[0] = NULL;
-    K_buf_[1] = NULL;
-
-    for(int i = 0; i < 2; ++i) {
-        for(int j = 0; j < label_J_[i].size(); ++j) {
-            delete[] label_J_[i][j];
-        }
-        for(int j = 0; j < label_K_[i].size(); ++j) {
-            delete[] label_K_[i][j];
-        }
-        label_J_[i].clear();
-        label_K_[i].clear();
-    }
-
-}
-
-void PK_integrals::deallocate_buffers() {
-    if(!in_core_) {
-        timer_on("AIO synchronize");
-        AIO_->synchronize();
-        timer_off("AIO synchronize");
-    }
-    // Can get rid of the pre-striping labels
-    for(int i = 0; i < label_J_[0].size(); ++i ) {
-        delete [] label_J_[0][i];
-    }
-    label_J_[0].clear();
-    for(int i = 0; i < label_K_[0].size(); ++i ) {
-        delete [] label_K_[0][i];
-    }
-    label_K_[0].clear();
-    for(int i = 0; i < nthreads_; ++i) {
-        delete iobuffers_[i];
-    }
-}
-
-size_t PK_integrals_old::task_quartets() {
-    // Need iterator over AO shells here
-    AOShellCombinationsIterator shelliter(primary_,primary_,primary_,primary_);
-    size_t nsh_task = 0;
-    for(shelliter.first(); shelliter.is_done() == false; shelliter.next()) {
-        int P = shelliter.p();
-        int Q = shelliter.q();
-        int R = shelliter.r();
-        int S = shelliter.s();
-        if (!sieve_->shell_significant(P,Q,R,S)) continue;
-        // Get the lowest basis function indices in the shells
-
-        size_t lowi = primary_->shell_to_basis_function(P);
-        size_t lowj = primary_->shell_to_basis_function(Q);
-        size_t lowk = primary_->shell_to_basis_function(R);
-        size_t lowl = primary_->shell_to_basis_function(S);
-
-        size_t low_ijkl = INDEX4(lowi, lowj, lowk, lowl);
-        size_t low_ikjl = INDEX4(lowi, lowk, lowj, lowl);
-        size_t low_iljk = INDEX4(lowi, lowl, lowj, lowk);
-
-        size_t nbJ_lo = low_ijkl / buffer_size_;
-        size_t nbK1_lo = low_ikjl / buffer_size_;
-        size_t nbK2_lo = low_iljk / buffer_size_;
-
-        // Can we filter out this shell because all of its basis function
-        // indices are too high ?
-
-        if (nbJ_lo > bufidx_ && nbK1_lo > bufidx_ && nbK2_lo > bufidx_) {
-            continue;
-        }
-
-        int ni = primary_->shell(P).nfunction();
-        int nj = primary_->shell(Q).nfunction();
-        int nk = primary_->shell(R).nfunction();
-        int nl = primary_->shell(S).nfunction();
-
-        size_t hii = lowi + ni - 1;
-        size_t hij = lowj + nj - 1;
-        size_t hik = lowk + nk - 1;
-        size_t hil = lowl + nl - 1;
-
-        size_t hi_ijkl = INDEX4(hii, hij, hik, hil);
-        size_t hi_ikjl = INDEX4(hii, hik, hij, hil);
-        size_t hi_iljk = INDEX4(hii, hil, hij, hik);
-
-        size_t nbJ_hi = hi_ijkl / buffer_size_;
-        size_t nbK1_hi = hi_ikjl / buffer_size_;
-        size_t nbK2_hi = hi_iljk / buffer_size_;
-
-        // Can we filter out this shell because all of its basis function
-        // indices are too low ?
-
-        if (nbJ_hi < bufidx_ && nbK1_hi < bufidx_ && nbK2_hi < bufidx_) {
-            continue;
-        }
-
-        // Now we loop over unique basis function quartets in the shell quartet
-        AOIntegralsIterator bfiter = shelliter.integrals_iterator();
-        for(bfiter.first(); bfiter.is_done() == false; bfiter.next()) {
-            size_t i = bfiter.i();
-            size_t j = bfiter.j();
-            size_t k = bfiter.k();
-            size_t l = bfiter.l();
-
-            size_t ijkl = INDEX4(i,j,k,l);
-            size_t ikjl = INDEX4(i,k,j,l);
-            size_t iljk = INDEX4(i,l,j,k);
-
-            size_t nbJ = ijkl / buffer_size_;
-            size_t nbK1 = ikjl / buffer_size_;
-            size_t nbK2 = iljk / buffer_size_;
-
-            if (nbJ == bufidx_ || nbK1 == bufidx_ || nbK2 == bufidx_) {
-                buf_P.push_back(P);
-                buf_Q.push_back(Q);
-                buf_R.push_back(R);
-                buf_S.push_back(S);
-                ++nsh_task;
-                break;
-            }
-        }
-    }
-    return nsh_task;
-}
-
-// This function is thread-safe
-void PK_integrals_old::integrals_buffering(const double *buffer, int P, int Q, int R, int S) {
-
-    AOIntegralsIterator bfiter(primary_->shell(P), primary_->shell(Q), primary_->shell(R), primary_->shell(S));
-
-    for (bfiter.first(); bfiter.is_done() == false; bfiter.next()) {
-        int i = bfiter.i();
-        int j = bfiter.j();
-        int k = bfiter.k();
-        int l = bfiter.l();
-        size_t idx = bfiter.index();
-
-        double val = buffer[idx];
-        fill_values(val, i, j, k, l);
-    }
-}
-
-void PK_integrals::integrals_buffering(const double *buffer, unsigned int P,
-                                       unsigned int Q, unsigned int R, unsigned int S) {
-    int thread = 0;
-#ifdef _OPENMP
-    thread = omp_get_thread_num();
-#endif
-    AOIntegralsIterator bfiter(primary_->shell(P), primary_->shell(Q), primary_->shell(R), primary_->shell(S));
-
-    for (bfiter.first(); bfiter.is_done() == false; bfiter.next()) {
-        int i = bfiter.i();
-        int j = bfiter.j();
-        int k = bfiter.k();
-        int l = bfiter.l();
-        size_t idx = bfiter.index();
-
-        double val = buffer[idx];
-        iobuffers_[thread]->fill_values(val, i, j, k, l);
-    }
-
-}
-
-// This function is thread-safe
-void PK_integrals_old::fill_values(double val, size_t i, size_t j, size_t k, size_t l) {
-    size_t ijkl = INDEX4(i, j, k, l);
-    if (ijkl / buffer_size_ == bufidx_) {
-#pragma omp atomic
-        J_buf_[buf_][ijkl - offset_] += val;
-    }
-
-    size_t ikjl = INDEX4(i, k, j, l);
-    if(ikjl / buffer_size_ == bufidx_) {
-        if (i == k || j == l) {
-#pragma omp atomic
-            K_buf_[buf_][ikjl - offset_] += val;
-        } else {
-#pragma omp atomic
-            K_buf_[buf_][ikjl - offset_] += 0.5 * val;
-        }
-    }
-
-    if(i != j && k != l) {
-        size_t iljk = INDEX4(i, l, j, k);
-        if (iljk / buffer_size_ == bufidx_) {
-            if ( i == l || j == k) {
-#pragma omp atomic
-                K_buf_[buf_][iljk - offset_] += val;
-            } else {
-#pragma omp atomic
-                K_buf_[buf_][iljk - offset_] += 0.5 * val;
-            }
-        }
-    }
-}
-
-void PK_integrals_old::open_files(bool old) {
-    if (in_core_) {
-        return;
-    }
-    psio_->open(itap_J_, old ? PSIO_OPEN_OLD : PSIO_OPEN_NEW);
-    psio_->open(itap_K_, old ? PSIO_OPEN_OLD : PSIO_OPEN_NEW);
-}
-
-void PK_integrals::open_files(bool old) {
-    if(in_core_) {
-        return;
-    }
-    psio_->open(pk_file_, old ? PSIO_OPEN_OLD : PSIO_OPEN_NEW);
-    // Create the files ? Pre-stripe them.
-    if(!old) {
-        for(int batch = 0; batch < batch_index_min_.size(); ++batch) {
-            size_t batch_size = batch_index_max_[batch] - batch_index_min_[batch];
-            // We need to keep the labels around in a vector
-            label_J_[0].push_back(get_label_J(batch));
-            AIO_->zero_disk(pk_file_,label_J_[0][batch],1,batch_size);
-            label_K_[0].push_back(get_label_K(batch));
-            AIO_->zero_disk(pk_file_,label_K_[0][batch],1,batch_size);
-        }
-
-    }
-}
-
-void PK_integrals::open_iwlf(bool old) {
-    if (in_core_) {
-        return;
-    }
-    psio_->open(iwl_file_J_, old ? PSIO_OPEN_OLD : PSIO_OPEN_NEW);
-    // Create the file ? Pre-stripe it
-    // Number of IWL buffers necessary
-    size_t num_iwlbuf = pk_size_ / ints_per_buf_ + 1;
-    // May need one more buffer at end of each batch
-    num_iwlbuf += batch_index_min_.size();
-    size_t iwlsize_bytes = num_iwlbuf * iwlintsize_;
-    size_t iwlsize = iwlsize_bytes / sizeof(double) + 1;
-    AIO_->zero_disk(iwl_file_J_,IWL_KEY_BUF,1,iwlsize);
-
-    // And we do the same for the file containing K buckets
-    // integrals can potentially go in 2 buckets for K
-    psio_->open(iwl_file_K_, old ? PSIO_OPEN_OLD : PSIO_OPEN_NEW);
-    AIO_->zero_disk(iwl_file_K_,IWL_KEY_BUF,1,2 * iwlsize);
-}
-
-char* PK_integrals_old::get_label_J(size_t i) {
-    char* label = new char[100];
-    sprintf(label, "J Block (Batch %d)", i);
-    return label;
-}
-char* PK_integrals_old::get_label_K(size_t i) {
-    char* label = new char[100];
-    sprintf(label, "K Block (Batch %d)", i);
-    return label;
-}
-
-// This function is not designed thread-safe and is supposed to execute outside parallel
-// environment for now.
-void PK_integrals_old::write() {
-    if (in_core_) {
-        return;
-    }
-    // Clear up the buffers holding P, Q, R, S since the task is over
-    buf_P.clear();
-    buf_Q.clear();
-    buf_R.clear();
-    buf_S.clear();
-    // Compute initial ijkl index for current buffer
-    size_t ijkl0 = buffer_size_ * bufidx_;
-    size_t ijkl_end = ijkl0 + buffer_size_ - 1;
-
-    std::vector<int> target_batches;  // Vector of batch number to which we want to write
-
-    for (int i = 0; i < batch_index_min_.size(); ++i) {
-        if (ijkl0 >= batch_index_min_[i] && ijkl0 < batch_index_max_[i]) {
-            target_batches.push_back(i);
-            continue;
-        }
-        if (ijkl_end >= batch_index_min_[i] && ijkl_end < batch_index_max_[i]) {
-            target_batches.push_back(i);
-            continue;
-        }
-        if (ijkl0 < batch_index_min_[i] && ijkl_end >= batch_index_max_[i]) {
-            target_batches.push_back(i);
-            continue;
-        }
-    }
-
-    // Now that all buffers are full, and before we write integrals, we need
-    // to divide diagonal elements by 2.
-
-    for (size_t pq = 0; pq < pk_pairs_; ++pq) {
-        size_t pqpq = INDEX2(pq, pq);
-        if (pqpq >= ijkl0 && pqpq <= ijkl_end) {
-            J_buf_[buf_][pqpq - offset_] *= 0.5;
-            K_buf_[buf_][pqpq - offset_] *= 0.5;
-        }
-    }
-
-    // And now we write to the files in the appropriate entries
-    for (int i = 0; i < target_batches.size(); ++i) {
-        char* label_J = new char[100];
-        sprintf(label_J,"J Block (Batch %d)", target_batches[i]);
-        label_J_[buf_].push_back(label_J);
-        size_t start = std::max(ijkl0, batch_index_min_[target_batches[i]]);
-        size_t stop = std::min(ijkl_end + 1, batch_index_max_[target_batches[i]]);
-        psio_address adr = psio_get_address(PSIO_ZERO, (start - batch_index_min_[target_batches[i]]) * sizeof(double));
-        size_t nints = stop - start;
-        jobid_J_[buf_] = AIO_->write(itap_J_, label_J_[buf_][i], (char *)(&J_buf_[buf_][start - offset_]),
-                    nints * sizeof(double), adr, &dummy_);
-        char* label_K = new char[100];
-        sprintf(label_K,"K Block (Batch %d)",target_batches[i]);
-        label_K_[buf_].push_back(label_K);
-        jobid_K_[buf_] = AIO_->write(itap_K_, label_K_[buf_][i], (char *)(&K_buf_[buf_][start - offset_]),
-                    nints * sizeof(double), adr, &dummy_);
-    }
-
-    // Update the buffer being written into
-    buf_ = buf_ == 0 ? 1 : 0;
-    // Make sure the buffer has been written to disk and we can erase it
-    AIO_->wait_for_job(jobid_J_[buf_]);
-    AIO_->wait_for_job(jobid_K_[buf_]);
-    // We can delete the labels for these buffers
-    for(int i = 0; i < label_J_[buf_].size(); ++i) {
-        delete [] label_J_[buf_][i];
-    }
-    for(int i = 0; i < label_K_[buf_].size(); ++i) {
-        delete [] label_K_[buf_][i];
-    }
-    label_J_[buf_].clear();
-    label_K_[buf_].clear();
-    // Make sure the buffer in which we are going to write is set to zero
-    ::memset((void *) J_buf_[buf_], '\0', buffer_size_ * sizeof(double));
-    ::memset((void *) K_buf_[buf_], '\0', buffer_size_ * sizeof(double));
-    // Update the offset for integral storage
-    offset_ += buffer_size_;
-    // And we are now writing to the next set of buffers
-    ++bufidx_;
-
-}
-
-// Thread-safe function, invokes the appropriate buffer to perform the write
-void PK_integrals::write() {
-    int thread = 0;
-#ifdef _OPENMP
-    thread = omp_get_thread_num();
-#endif
-    iobuffers_[thread]->write(batch_index_min_,batch_index_max_,pk_pairs_);
-}
-
-void PK_integrals::write_iwl() {
-    // Each thread buffer has several buffers that may be partially full.
-    // Concatenate all internal buffers to the ones of thread buffer 0.
-    double val;
-    size_t i, j, k, l;
-    IOBuffer_PK* buf0 = iobuffers_[0];
-    IOBuffer_PK* buftarget;
-    //TODO Man if that loop actually works I'll be lucky
-    for(int t = 1; t < nthreads_; ++t) {
-        buftarget = iobuffers_[t];
-        unsigned int nbufs = buftarget->nbuf();
-        // Factor 2 to get buffers for J and K
-        for(int buf = 0; buf < 2 * nbufs; ++buf) {
-            while(buftarget->pop_value(buf,val,i,j,k,l)) {
-                buf0->insert_value(buf,val,i,j,k,l);
-            }
-        }
-
-    }
-    // By now we should have transferred all remaining integrals to
-    // buffer 0, we now flush it.
-    buf0->flush();
-}
-
-void PK_integrals_old::close_files() {
-    if(in_core_) {
-        return;
-    }
-    psio_->close(itap_J_, 1);
-    psio_->close(itap_K_, 1);
-}
-
-void PK_integrals::close_files() {
-    if(in_core_) {
-        return;
-    }
-    psio_->close(pk_file_, 1);
-}
-
-void PK_integrals::close_iwlf(bool del) {
-    if(in_core_) {
-        return;
-    }
-    psio_->close(iwl_file_J_, del ? 0 : 1);
-    psio_->close(iwl_file_K_, del ? 0 : 1);
-}
-
-void PK_integrals::sort_ints() {
-    // compute max batch size
-    size_t max_size = 0;
-    size_t batch_size;
-    int nbatches = batch_index_min_.size();
-
-    for(int i = 0; i < nbatches; ++i) {
-        batch_size = batch_index_max_[i] - batch_index_min_[i];
-        if(batch_size > max_size) max_size = batch_size;
-    }
-
-    // Allocate a big array to contain the batch.
-    double* twoel_ints = new double[max_size];
-    ::memset((void*)twoel_ints,'\0',max_size * sizeof(double));
-
-    // At this point we need to close the IWL file to create
-    // an IWL object which will open it. Dumb but minor inconvenience (hopefully)
-    // We also open the PK file
-    psio_->open(pk_file_, PSIO_OPEN_NEW);
-    deallocate_buffers();
-    set_writing(false);
-    close_iwlf(false);
-    // Probably here, a call to a function for sorting
-
-    generate_J_PK(twoel_ints,max_size);
-    // Need to reset to zero the two-el integral array
-    ::memset((void*)twoel_ints,'\0',max_size * sizeof(double));
-    generate_K_PK(twoel_ints,max_size);
-
-    psio_->close(pk_file_,1);
-
-}
-
-void PK_integrals::generate_J_PK(double *twoel_ints, size_t max_size) {
-
-    IWL inbuf(psio_.get(),iwl_file_J_,0.0, 1, 0);
-
-    int idx;
-    size_t p, q, r, s;
-    Label* lblptr = inbuf.labels();
-    Value* valptr = inbuf.values();
-    int lastbuf;
-    double val;
-
-    size_t offset, maxind, nintegrals;
-    size_t pqrs;
-
-    int batch = 0;
-    int nbatches = batch_index_min_.size();
-    while(batch < nbatches) {
-        inbuf.fetch();
-        offset = batch_index_min_[batch];
-        maxind = batch_index_max_[batch];
-        nintegrals = batch_index_max_[batch] - batch_index_min_[batch];
-
-        for(idx = 0; idx < inbuf.buffer_count(); ++idx) {
-            p = lblptr[4 * idx];
-            q = lblptr[4 * idx + 1];
-            r = lblptr[4 * idx + 2];
-            s = lblptr[4 * idx + 3];
-
-            pqrs = INDEX4(p,q,r,s);
-            //TODO: remove the check, for debug only
-            if(pqrs < offset || pqrs > maxind) {
-                throw PSIEXCEPTION("Integrals not written properly \n");
-            }
-            twoel_ints[pqrs - offset] += valptr[idx];
-        }
-
-        lastbuf = inbuf.last_buffer();
-        if(lastbuf) {
-            // We just completed a batch, write it to disk
-            char* label = get_label_J(batch);
-            // Divide by two diagonal elements
-            for(size_t pq = batch_pq_min_[batch]; pq < batch_pq_max_[batch]; ++pq) {
-                pqrs = INDEX2(pq,pq);
-                twoel_ints[pqrs - offset] *= 0.5;
-            }
-            psio_->write_entry(pk_file_, label, (char*)twoel_ints, nintegrals * sizeof(double));
-            ++batch;
-            if(batch < nbatches) {
-                ::memset((void*)twoel_ints,'\0',max_size * sizeof(double));
-            }
-        }
-
-    }
-
-    inbuf.set_keep_flag(false);
-
-}
-
-void PK_integrals::generate_K_PK(double *twoel_ints, size_t max_size) {
-
-    IWL inbuf(psio_.get(),iwl_file_K_,0.0, 1, 0);
-
-    int idx;
-    size_t p, q, r, s;
-    Label* lblptr = inbuf.labels();
-    Value* valptr = inbuf.values();
-    int lastbuf;
-    double val;
-
-    size_t offset, maxind, nintegrals;
-    size_t pqrs;
-
-    int batch = 0;
-    int nbatches = batch_index_min_.size();
-    while(batch < nbatches) {
-        inbuf.fetch();
-        offset = batch_index_min_[batch];
-        maxind = batch_index_max_[batch];
-        nintegrals = batch_index_max_[batch] - batch_index_min_[batch];
-
-        for(idx = 0; idx < inbuf.buffer_count(); ++idx) {
-            p = lblptr[4 * idx];
-            q = lblptr[4 * idx + 1];
-            r = lblptr[4 * idx + 2];
-            s = lblptr[4 * idx + 3];
-
-
-            // K first sort
-            pqrs = INDEX4(p,r,q,s);
-            size_t pqd = INDEX2(p,r);
-            size_t rsd = INDEX2(q,s);
-            if(pqrs <= maxind && pqrs >= offset) {
-                if(p == r || q == s) {
-                    twoel_ints[pqrs - offset] += valptr[idx];
-                } else {
-                    twoel_ints[pqrs - offset] += 0.5 * valptr[idx];
-                }
-//DEBUG                if(pqd == 0 && rsd == 0) {
-//DEBUG                    outfile->Printf("Int is %20.16f\n",twoel_ints[pqrs - offset]);
-//DEBUG                }
-            }
-
-            // K second sort
-            if(p != q && r != s) {
-                pqrs = INDEX4(p,s,q,r);
-                if(pqrs <= maxind && pqrs >= offset) {
-                    if(p == s || q == r) {
-                        twoel_ints[pqrs - offset] += valptr[idx];
-                    } else {
-                        twoel_ints[pqrs - offset] += 0.5 * valptr[idx];
-                    }
-//DEBUG                if(pqd == 0 && rsd == 0) {
-//DEBUG                    outfile->Printf("Int is %20.16f\n",twoel_ints[pqrs - offset]);
-//DEBUG                }
-                }
-
-            }
-        }
-
-        lastbuf = inbuf.last_buffer();
-        if(lastbuf) {
-            // We just completed a batch, write it to disk
-            char* label = get_label_K(batch);
-            // Divide by two diagonal elements
-            for(size_t pq = batch_pq_min_[batch]; pq < batch_pq_max_[batch]; ++pq) {
-                pqrs = INDEX2(pq,pq);
-                twoel_ints[pqrs - offset] *= 0.5;
-            }
-            psio_->write_entry(pk_file_, label, (char*)twoel_ints, nintegrals * sizeof(double));
-            ++batch;
-            if(batch < nbatches) {
-                ::memset((void*)twoel_ints,'\0',max_size * sizeof(double));
-            }
-        }
-
-    }
-
-    inbuf.set_keep_flag(false);
-
-}
-
-// We should really use safer pointers right here
-void PK_integrals_old::form_D_vec(std::vector<SharedMatrix> D_ao) {
-    // Assume symmetric density matrix for now.
-    for (int N = 0; N < D_ao.size(); ++N) {
-        double* D_vec = new double[pk_pairs_];
-        ::memset((void *)D_vec,'\0',pk_pairs_ * sizeof(double));
-        D_vec_.push_back(D_vec);
-        size_t pqval = 0;
-        for(int p = 0; p < nbf_; ++p) {
-            for(int q = 0; q <= p; ++q) {
-                if(p != q) {
-                    D_vec[pqval] = 2.0 * D_ao[N]->get(0,p,q);
-                } else {
-                    D_vec[pqval] = D_ao[N]->get(0,p,q);
-                }
-                ++pqval;
-            }
-        }
-    }
-}
-
-void PK_integrals_old::finalize_D() {
-    for(int N = 0; N < D_vec_.size(); ++N) {
-        delete [] D_vec_[N];
-    }
-    D_vec_.clear();
-}
-
-void PK_integrals_old::form_J(std::vector<SharedMatrix> J, bool exch) {
-    std::vector<double*> Jvecs;
-    // Begin by allocating vector for triangular J
-    for(int N = 0; N < J.size(); ++N) {
-        double* J_vec = new double[pk_pairs_];
-        ::memset((void*)J_vec,'\0',pk_pairs_ * sizeof(double));
-        Jvecs.push_back(J_vec);
-    }
-
-    if(in_core_) {
-        // Read one entry, use it for all density matrices
-        for(int N = 0; N < J.size(); ++N) {
-            double* D_vec = D_vec_[N];
-            double* J_vec = Jvecs[N];
-            double* j_ptr;
-            if(exch) {
-                j_ptr = K_ints_;
-            } else {
-                j_ptr = J_ints_;
-            }
-//TODO We should totally parallelize this loop now.
-            for(size_t pq = 0; pq < pk_pairs_; ++pq) {
-                double D_pq = D_vec[pq];
-                double *D_rs = D_vec;
-                double J_pq = 0.0;
-                double *J_rs = J_vec;
-                for(size_t rs = 0; rs <= pq; ++rs) {
-//DEBUG                    if(exch) { // && rs == 0) {
-//DEBUG                      outfile->Printf("PK int (%lu|%lu) = %20.16f\n",pq,rs,*j_ptr);
-//DEBUG                    
-                    J_pq += *j_ptr * (*D_rs);
-                    *J_rs += *j_ptr * D_pq;
-                    ++D_rs;
-                    ++J_rs;
-                    ++j_ptr;
-                }
-                J_vec[pq] += J_pq;
-            }
-        }
-    } else {
-        // Now loop over batches
-        for(int batch = 0; batch < batch_pq_min_.size(); ++batch) {
-            size_t min_index = batch_index_min_[batch];
-            size_t max_index = batch_index_max_[batch];
-            size_t batch_size = max_index - min_index;
-            size_t min_pq = batch_pq_min_[batch];
-            size_t max_pq = batch_pq_max_[batch];
-            double* j_block = new double[batch_size];
-
-            int filenum;
-
-            char* label;
-            if (exch) {
-                label = get_label_K(batch);
-                filenum = itap_K_;
-            } else {
-                label = get_label_J(batch);
-                filenum = itap_J_;
-            }
-            psio_->read_entry(filenum, label, (char *) j_block, batch_size * sizeof(double));
-
-            // Read one entry, use it for all density matrices
-            for(int N = 0; N < J.size(); ++N) {
-                double* D_vec = D_vec_[N];
-                double* J_vec = Jvecs[N];
-                double* j_ptr = j_block;
-            //TODO Could consider parallelizing this loop
-                for(size_t pq = min_pq; pq < max_pq; ++pq) {
-                    double D_pq = D_vec[pq];
-                    double *D_rs = D_vec;
-                    double J_pq = 0.0;
-                    double *J_rs = J_vec;
-                    for(size_t rs = 0; rs <= pq; ++rs) {
-//DEBUG                        if(exch) { // && rs == 0) {
-//DEBUG                          outfile->Printf("PK int (%lu|%lu) = %20.16f\n",pq,rs,*j_ptr);
-//DEBUG                        }
-                        J_pq += *j_ptr * (*D_rs);
-                        *J_rs += *j_ptr * D_pq;
-                        ++D_rs;
-                        ++J_rs;
-                        ++j_ptr;
-                    }
-                    J_vec[pq] += J_pq;
-                }
-            }
-
-            delete [] label;
-            delete [] j_block;
-        }
-    }
-
-    // Now, directly transfer data to resulting matrices
-    for(int N = 0; N < J.size(); ++N) {
-        double *Jp = Jvecs[N];
-        for(int p = 0; p < nbf_; ++p) {
-            for(int q = 0; q <= p; ++q) {
-                J[N]->set(0,p,q,*Jp++);
-            }
-        }
-        J[N]->copy_lower_to_upper();
-        delete [] Jvecs[N];
-    }
-}
-
-void PK_integrals_old::form_K(std::vector<SharedMatrix> K) {
-    // For symmetric densities, we do exactly the same than for J
-    // but we read another entry
-    form_J(K, true);
-}
 
 }
